+++ src/main/java/wooteco/subway/domain/line/LineDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.station.dao;
+package wooteco.subway.domain.line;
 
 import java.sql.PreparedStatement;
 import java.util.List;
@@ -10,30 +10,31 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.station.Station;
 
 @Repository
-public class StationDao {
+public class LineDao {
 
     private final JdbcTemplate jdbcTemplate;
 
-    private final RowMapper<Station> stationRowMapper = (rs, rowNum) -> new Station(
+    private final RowMapper<Line> lineRowMapper = (rs, rowNum) -> new Line(
             rs.getLong("id"),
-            rs.getString("name")
+            rs.getString("name"),
+            rs.getString("color")
     );
 
-    public StationDao(JdbcTemplate jdbcTemplate) {
+    public LineDao(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Optional<Long> save(Station station) {
+    public Optional<Long> save(Line line) {
         final KeyHolder keyHolder = new GeneratedKeyHolder();
-        final String sql = "INSERT INTO station (name) VALUES (?)";
+        final String sql = "INSERT INTO line (name, color) VALUES (?, ?)";
 
         try {
             jdbcTemplate.update(con -> {
                 PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"});
-                ps.setString(1, station.getName());
+                ps.setString(1, line.getName());
+                ps.setString(2, line.getColor());
                 return ps;
             }, keyHolder);
         } catch (DuplicateKeyException e) {
@@ -44,24 +45,28 @@ public class StationDao {
         return Optional.of(keyHolder.getKey().longValue());
     }
 
-    public List<Station> findAll() {
-        final String sql = "SELECT id, name FROM station";
-        return jdbcTemplate.query(sql, stationRowMapper);
+    public List<Line> findAll() {
+        final String sql = "SELECT id, name, color FROM line";
+        return jdbcTemplate.query(sql, lineRowMapper);
     }
 
-    public Optional<Station> findById(Long id) {
-        final String sql = "SELECT id, name FROM station WHERE id = ?";
-
+    public Optional<Line> findById(Long id) {
+        final String sql = "SELECT id, name, color FROM line WHERE id = ?";
         try {
-            final Station station = jdbcTemplate.queryForObject(sql, stationRowMapper, id);
-            return Optional.of(station);
+            Line line = jdbcTemplate.queryForObject(sql, lineRowMapper, id);
+            return Optional.of(line);
         } catch (EmptyResultDataAccessException e) {
             return Optional.empty();
         }
     }
 
+    public void update(Long id, Line line) {
+        final String sql = "UPDATE line SET name = ?, color = ? WHERE id = ?";
+        jdbcTemplate.update(sql, line.getName(), line.getColor(), id);
+    }
+
     public void delete(Long id) {
-        final String sql = "DELETE FROM station WHERE id = ?";
+        final String sql = "DELETE FROM line WHERE id = ?";
         jdbcTemplate.update(sql, id);
     }
 }
