+++ src/main/java/wooteco/subway/service/LineService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.domain.line.dao.LineDao;
+import wooteco.subway.domain.line.LineDao;
 import wooteco.subway.web.exception.NotFoundException;
 import wooteco.subway.web.exception.SubwayHttpException;
